remove cast warnings (-Wconversion)

+++ blkio.h
@@ -19,7 +19,7 @@ struct wring;
 
 // iosz: fixed write size; must be a multiple of PGSZ
   extern struct wring *
-wring_create(const int fd, const size_t iosz, const u32 depth);
+wring_create(const int fd, const u32 iosz, const u32 depth);
 
   extern void
 wring_update_fd(struct wring * const wring, const int fd);
@@ -32,11 +32,11 @@ wring_acquire(struct wring * const wring);
 
 // write part of the buf
   extern void
-wring_write_partial(struct wring * const wring, const size_t off,
-    void * const buf, const size_t buf_off, const size_t size);
+wring_write_partial(struct wring * const wring, const off_t off,
+    void * const buf, const size_t buf_off, const u32 size);
 
   extern void
-wring_write(struct wring * const wring, const size_t off, void * const buf);
+wring_write(struct wring * const wring, const off_t off, void * const buf);
 
 // flush the queue and wait for completion
   extern void
